feat(nuxt): add `onNuxtReady` composable
